Экран подключен во вкладку таббара профиля,изменен метод для создания сцены в Factory, верстка вынесена из VC в отдельный файл.
