+++ iaw01/AppDelegate/RootTabBar/RootTabBarController.swift
@@ -25,7 +25,7 @@ final class RootTabBarController: UITabBarController {
             configureController(with: factory.createDummyScene(), tabBarItem: .discover),
             configureController(with: factory.createDummyScene(), tabBarItem: .drivethru),
             configureController(with: factory.createDummyScene(), tabBarItem: .orders),
-            configureController(with: factory.createDummyScene(), tabBarItem: .profile),
+            configureController(with: factory.createVerifyPhoneNumberViewController(), tabBarItem: .profile),
         ]
     }
